app/pyproject.toml
[project]
name = "chess-llm-ui"
version = "0.1.0"
description = "Play chess against a fine-tuned LLM with a Python-based UI"
authors = [{ name = "Hridaya Sharma" }]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "transformers>=4.44.0",
    "torch>=2.3.0",
    "python-chess>=1.999",
    "pyqt5>=5.15.9",
    "cairosvg>=2.8.2",
    "accelerate>=1.10.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
python-preference = "system"

